% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aCGH}
\alias{aCGH}
\title{Transformed aCGH data}
\format{
A data frame with 2046 rows and 1 variable:
\describe{
\item{transNorm}{normalized aCGH intensity}
}
}
\usage{
aCGH
}
\description{
A dataset containing the tranformed aCGH data from the genome of the fibroblast cell line GM02948
}
\keyword{datasets}
