% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_cp.R
\name{addDatapoint}
\alias{addDatapoint}
\title{Add one data point}
\usage{
addDatapoint(bcp, d)
}
\arguments{
\item{bcp, }{current BayesCP object}

\item{d, }{additional data point to be added to the existing data}
}
\value{
a vector with new data point appended
}
\description{
Add one data point
}
