% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnv_H2347}
\alias{cnv_H2347}
\title{GC-corrected data for copy number variation}
\format{
A data frame with 14189 rows and 2 variables:
\describe{
  \item{raw.count}{raw read counts}
  \item{normalized.count}{normalized read counts}
}
}
\usage{
cnv_H2347
}
\description{
A dataset containing the raw data and GC-corrected/normalized data
}
\keyword{datasets}
