% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid}
\alias{covid}
\title{US COVID-19 data}
\format{
A data frame with 219 rows and 8 variables
\describe{
\item{Date_reported}{The report date}
\item{Country_code}{The code for country}
\item{Country}{Country in full name}
\item{WHO_region}{Geographic region defined by WHO}
\item{New_cases}{New COVID-19 cases}
\item{Cumulative_cases}{Cumulative COVID-19 cases}
\item{New_deaths}{New COVID-19 deaths}
\item{Cumulative_deaths}{Cumulative COVID-19 deaths}
}
}
\usage{
covid
}
\description{
A dataset containing new daily cases in the United States downloaded from the
World Health Organization on August 25, 2020
}
\keyword{datasets}
