% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_cp.R
\name{imputation}
\alias{imputation}
\title{Impute missing data}
\usage{
imputation(x, method = c("Median", "kNN"))
}
\arguments{
\item{x}{the normalized data with missing}

\item{method}{the imputation method}
}
\value{
The vector of imputed data with no missing values
}
\description{
Impute missing data
}
