% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_cp.R
\name{summary.BayesCP}
\alias{summary.BayesCP}
\title{Summarize BayesCP object}
\usage{
\method{summary}{BayesCP}(object, norm.test = FALSE, ...)
}
\arguments{
\item{object}{the BayesCP class object to be summarized}

\item{norm.test}{logical value for normality test, default is false}

\item{...}{parameters passed to summary()}
}
\value{
The summary of the BayesCP class in a list
}
\description{
Summarize BayesCP object
}
\examples{
x <- c(rnorm(10, 0, 1), rnorm(10, 5, 1))
bcp <- online_cp(x)
summary(bcp)
}
