% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_cp.R
\name{combine}
\alias{combine}
\title{Combine two BayesCP objects}
\usage{
combine(bcp1, bcp2)
}
\arguments{
\item{bcp1}{the first BayesCP object to be combined}

\item{bcp2}{the second BayesCP opbject to be combined}
}
\value{
The combined BayesCP object
}
\description{
Combine two BayesCP objects
}
