% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_cp.R
\name{plot.BayesCP}
\alias{plot.BayesCP}
\title{Plot BayesCP object}
\usage{
\method{plot}{BayesCP}(x, xlab = "Index", ylab = "x", ...)
}
\arguments{
\item{x}{the BayesCP class object to be plotted}

\item{xlab}{the default x-axis label, default "Index"}

\item{ylab}{the default y-axis label, default "x"}

\item{...}{the plotting parameters passed to plot()}
}
\value{
No return value, called for side effects
}
\description{
Plot BayesCP object
}
